Add token update function to System Log. Ref #86

+++ backend/src/main/java/org/hfut/tool/global/Token.java
@@ -20,7 +20,7 @@ public class Token {
 
     private static final String SECRET = "*>*$)$_NOT_FOUND*>*";
 
-    public static String getToken(Integer projectPeopleId, Integer permissions) {
+    public static String getToken(Integer projectPeopleId, Byte permissions) {
         Date iatDate = new Date();
         Date expiresDate = new Date(iatDate.getTime() + (long) 1000 * 60 * 60 * 12);
 
@@ -28,7 +28,10 @@ public class Token {
         map.put("alg", "HS256");
         map.put("typ", "Token");
 
-        String token = JWT.create().withHeader(map).withClaim("iss", "org.htut.404NotFound").withClaim("aud", "EDMS_APP").withClaim("projectPeopleId", projectPeopleId).withClaim("permissions", permissions).withIssuedAt(iatDate).withExpiresAt(expiresDate).sign(Algorithm.HMAC256(SECRET));
+        String token = JWT.create().withHeader(map).withClaim("iss", "org.htut.404NotFound")
+                .withClaim("aud", "EDMS_APP").withClaim("projectPeopleId", projectPeopleId)
+                .withClaim("permissions", permissions.intValue()).withIssuedAt(iatDate).withExpiresAt(expiresDate)
+                .sign(Algorithm.HMAC256(SECRET));
         return token;
     }
